+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Rename all occurences of Kizzy to Axe
 - Experimental button shapes for the main screen
 - Developer credits in credits page
+- Modern rounded text input fields in custom RPC
+- Experimental note for button shapes in display settings
 ### Fixed
 - Timezone offset issue in custom rpc timestamp picker. ([rootdevss pull request on Kizzy](https://github.com/dead8309/Kizzy/pull/480/changes/f1106546e3dea3135929e4e372488654cb9dc927))
 - Corrected Turkish translations for clarity and consistency. ([kyoyacchi's pull request on Kizzy](https://github.com/dead8309/Kizzy/pull/491))